requirements.txt
deta==1.1.0
plotly==5.9.0
python-dotenv==1.0.0
streamlit_authenticator==0.2.2
streamlit_option_menu==0.3.6
